+++ lib/swift_animations.dart
@@ -21,4 +21,5 @@
 library;
 
 export 'src/animate.dart';
+export 'src/gesture_extensions.dart';